Fixed some value setting issues, added oper, updated unit test to ingest yang
