+++ tests/unit_test/test_generators.py
@@ -87,7 +87,7 @@ class BaseTest(unittest.TestCase):
 class TestCreateSql(BaseTest):
 
     def test_create_sql(self):
-        adm = self._get_adm('test_adm.json')
+        adm = self._get_adm('test-adm.yang')
         outdir = os.path.join(os.environ['XDG_DATA_HOME'], 'out')
 
         writer = create_sql.Writer(self._admset, adm, outdir, dialect='pgsql')
@@ -107,7 +107,7 @@ class TestCreateSql(BaseTest):
 class TestCreateCH(BaseTest):
 
     def test_create_impl_h_noscrape(self):
-        adm = self._get_adm('test_adm.json')
+        adm = self._get_adm('test-adm.yang')
         outdir = os.path.join(os.environ['XDG_DATA_HOME'], 'out')
 
         writer = create_impl_h.Writer(self._admset, adm, outdir, H_Scraper(None))
@@ -124,7 +124,7 @@ class TestCreateCH(BaseTest):
         self.assertEqual(content, buf.getvalue())
 
     def test_create_impl_c_noscrape(self):
-        adm = self._get_adm('test_adm.json')
+        adm = self._get_adm('test-adm.yang')
         outdir = os.path.join(os.environ['XDG_DATA_HOME'], 'out')
 
         writer = create_impl_c.Writer(self._admset, adm, outdir, C_Scraper(None))
@@ -140,42 +140,8 @@ class TestCreateCH(BaseTest):
         content = tmpl.render(datestamp=self._today_datestamp())
         self.assertEqual(content, buf.getvalue())
 
-    def test_create_gen_h(self):
-        adm = self._get_adm('test_adm.json')
-        outdir = os.path.join(os.environ['XDG_DATA_HOME'], 'out')
-
-        writer = create_gen_h.Writer(self._admset, adm, outdir)
-        self.assertEqual(
-            os.path.join(outdir, 'shared', 'adm', 'adm_test_adm.h'),
-            writer.file_path()
-        )
-
-        buf = io.StringIO()
-        writer.write(buf)
-
-        tmpl = self._tmpl_env.get_template('gen_ch/shared/adm/adm_test_adm.h.jinja')
-        content = tmpl.render(datestamp=self._today_datestamp())
-        self.assertEqual(content, buf.getvalue())
-
-    def test_create_mgr_c(self):
-        adm = self._get_adm('test_adm.json')
-        outdir = os.path.join(os.environ['XDG_DATA_HOME'], 'out')
-
-        writer = create_mgr_c.Writer(self._admset, adm, outdir)
-        self.assertEqual(
-            os.path.join(outdir, 'mgr', 'adm_test_adm_mgr.c'),
-            writer.file_path()
-        )
-
-        buf = io.StringIO()
-        writer.write(buf)
-
-        tmpl = self._tmpl_env.get_template('gen_ch/mgr/adm_test_adm_mgr.c.jinja')
-        content = tmpl.render(datestamp=self._today_datestamp())
-        self.assertEqual(content, buf.getvalue())
-
     def test_create_agent_c(self):
-        adm = self._get_adm('test_adm.json')
+        adm = self._get_adm('test-adm.yang')
         outdir = os.path.join(os.environ['XDG_DATA_HOME'], 'out')
 
         writer = create_agent_c.Writer(self._admset, adm, outdir)
